Makefile changes for static libs and HP-UX,   Compile fixes for HP-UX,
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@3904 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk1/dcclient.cpp
@@ -211,7 +211,7 @@ void wxWindowDC::DoDrawArc( long x1, long y1, long x2, long y2,
     long yyc = YLOG2DEV(yc);
     double dx = xx1 - xxc;
     double dy = yy1 - yyc;
-    double radius = sqrt(dx*dx+dy*dy);
+    double radius = sqrt((double)(dx*dx+dy*dy));
     long   r      = (long)radius;
     double radius1, radius2;
